improved Dockerfile stuff kinda
So now we have the rpmbuild container (which makes a repo out of its
inputs), and the testenv container (which copies in a repo from rpms/
and sets it up in DNF, and that's it)

The Makefile pulls the RPMs out of rpmbuild and crams them into testenv,
which isn't particularly elegant but I'm still not totally sure how to
make VOLUME work the way I want here.

Anyway, this at least gets us a container with our freshly-built code
ready to be installed, and that's what we need to start testing. Yay!