spec: Require package containing F23 keys
Attempted upgrades to F23 fail if this package isn't installed, and the
error message isn't particularly clear.

IMHO the simplest thing for F21 is just to make this package require the
necessary package and be done with it.

I think there's gonna have to be a smarter way for us to declare
packages that are needed by an upgrade from a particular version to
another particular version, but I'm not sure how best to do that yet.

So, in the meantime, add the Requires here and let's move on.